+++ CHANGELOG.rst
@@ -15,9 +15,11 @@ All notable changes to this project are documented in this file.
 - Fix multi-signature contract import to allow using a single signature
 - Fix fund sending from multi-signature contract
 - Added instructions on retrieving NEO TestNet funds
-- prompt.py: When using a privnet with ``-p``, check if chain database is correct. Renamed ``Chains/Priv_Notif`` to ``Chains/privnet_notif`` (if you need your old privnet notification db, you need to rename it manually).
-- Fixed issue with missing ``addr`` location in ``neo/api/REST/RestApi.py``
+- Fixed issue with missing ``notifications/`` prefix for ``addr`` call in ``neo/api/REST/RestApi.py``
 - Added ``neo-privnet.wallet`` to the project root. This is the standard wallet for `private networks <https://hub.docker.com/r/cityofzion/neo-privatenet/>`_.
+- prompt.py: When using a privnet with ``-p``, check if chain database is correct. Renamed ``Chains/Priv_Notif`` to ``Chains/privnet_notif`` (if you need your old privnet notification db, you need to rename it manually).
+- Optionally allow to use custom privnet hosts with ``-p`` (`PR #312 <https://github.com/CityOfZion/neo-python/pull/312>`_)
+- Added a dependency check to ``Settings.py``, which verifies that the installed dependencies match those in requirements.txt
 
 
 [0.5.3] 2018-03-04