+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ All notable changes to this project are documented in this file.
 - Added VM instruction counter to ``ExecutionEngine.py`` error messages to indicate the final instruction that failed. Allows for setting conditional breakpoints to support SC debugging.
 - Renamed ``neo.api.REST.NotificationRestApi`` to ``neo.api.REST.RestApi``
 - Added ``-v/--verbose`` argument to prompt.py, which makes prompt.py show smart contract events by default
+- Added ``vm-log`` option to the ``config`` console command. This enabled logging of VM instructions to ``vm_instructions.log`` for debugging purposes.
 - Fix multi-signature contract import to allow using a single signature
 - Fix fund sending from multi-signature contract
 - Added instructions on retrieving NEO TestNet funds