+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 
 All notable changes to this project are documented in this file.
 
-[0.5.4-dev] Work in progress
+[0.5.4] 2018-03-14
 ----------------------------
 - All requests to the API that are invalid will now receive a ``None`` for results rather than an empty list ``[]``
 - update to neo-boa==0.3.7